+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+3.0.3 / 2015-08-13
+==================
+
+* Deprecation warnings for old plugin nebula-javadoc-jar, nebula-source-jar, nebula-test-jar
+* Bug fix for multiproject builds with project dependencies via Viacheslav Shvets (slavashvets)                                                                                            
+
 3.0.2 / 2015-08-12
 ==================